Add 5V LDO to block diagram
It is possible to power the I2C isolator, I2C GPIO expander and power
monitor IC from the 5V5 SMPS. However, if it turns off due to
over-current/temperature it will be impossible for the micro to detect.
By powering the 'logic' parts on the isolated side from a separate LDO
it will be possible to monitor the SMPS's PGOOD signal and read the
voltage/current even when it's in trouble.