Add user UART socket
With very few components it's possible to expose the UART via a
mini-camcon socket, like on the motor board. There's no reason to not do
this (other than the cost of the handful of components and slightly
increased assembly cost).

It might even make sense to remove the FTDI header and replace it with a
button to enter the bootloader. Then flashing a brand new board is the
same as a motor board (albeit by plugging into the UART socket rather
than USB).