Get rid of FTDI header and add reset/bootloader button
The FTDI header doesn't offer much over just using the UART socket to
program a brand-new board. The intention of the RTS line of the FTDI
cable being connected to the BOOT0 line was that one could cause the
micro to enter into the bootloader by setting the line high. However,
you also need to reset the micro. There was no way of resetting the
micro and the FTDI cable powered the board, so you couldn't power cycle
it.

The BOOT0+nReset circuit has already been proven on the motor controller
(MCv4B). Flashing a brand-new servo board is now very simiarl to a
brand-new MCv4B, with the exception that you can only do it over the
UART socket (it's possible to flash an MCv4B over both the UART socket
and the USB port).