Assign some values. Remove superfluous pull-ups
The pull-up for the PGOOD signal can be provided by the internal pull-up
of the GPIO expander. The SMPS_EN had a pull-up to ensure it was off
when the board was first powered. The FET was then to pull that line
down to turn the SMPS on. However, there's no reason why the SMPS_EN
signal can't go straight into the GPIO expander.